rename schema to types in every module
