+++ backend/dataall/core/cognito_groups/api/__init__.py
@@ -1,4 +1,3 @@
-from . import input_types, queries, resolvers, schema
+from . import input_types, queries, resolvers, types
 
-__all__ = ['resolvers', 'input_types', 'schema', 'queries']
-
+__all__ = ['resolvers', 'input_types', 'types', 'queries']
