+++ backend/dataall/core/environment/api/__init__.py
@@ -3,8 +3,8 @@ from . import (
     mutations,
     queries,
     resolvers,
-    schema,
     enums,
+    types,
 )
 
-__all__ = ['resolvers', 'schema', 'input_types', 'queries', 'mutations', "enums"]
+__all__ = ['resolvers', 'types', 'input_types', 'queries', 'mutations', "enums"]
