+++ backend/dataall/core/feed/api/__init__.py
@@ -1,1 +1,1 @@
-from dataall.core.feed.api import input_types, mutations, queries, registry, resolvers, schema
+from dataall.core.feed.api import input_types, mutations, queries, registry, resolvers, types
