+++ backend/dataall/core/permissions/api/__init__.py
@@ -2,8 +2,8 @@ from dataall.core.permissions.api import (
     input_types,
     queries,
     resolvers,
-    schema,
+    types,
     mutations
 )
 
-__all__ = ['resolvers', 'schema', 'input_types', 'queries', 'mutations']
+__all__ = ['resolvers', 'types', 'input_types', 'queries', 'mutations']
