+++ backend/dataall/core/vote/api/__init__.py
@@ -2,8 +2,8 @@ from . import (
     input_types,
     queries,
     resolvers,
-    schema,
     mutations,
+    types,
 )
 
-__all__ = ['resolvers', 'schema', 'input_types', 'queries', 'mutations']
+__all__ = ['resolvers', 'types', 'input_types', 'queries', 'mutations']
