+++ backend/dataall/core/vote/api/types.py
