+++ backend/dataall/core/vpc/api/__init__.py
@@ -1,4 +1,4 @@
-from . import input_types, queries, mutations, resolvers, schema
+from . import input_types, queries, mutations, resolvers, types
 
-__all__ = ['resolvers', 'schema', 'input_types', 'queries', 'mutations']
+__all__ = ['resolvers', 'types', 'input_types', 'queries', 'mutations']
