+++ backend/dataall/modules/datapipelines/api/__init__.py
@@ -3,7 +3,7 @@ from dataall.modules.datapipelines.api import (
     mutations,
     queries,
     resolvers,
-    schema,
+    types,
 )
 
-__all__ = ['resolvers', 'schema', 'input_types', 'queries', 'mutations']
+__all__ = ['resolvers', 'types', 'input_types', 'queries', 'mutations']
