+++ backend/dataall/modules/datasets/api/storage_location/mutations.py
@@ -8,7 +8,7 @@ from dataall.modules.datasets.api.storage_location.resolvers import (
     update_storage_location,
     remove_storage_location,
 )
-from dataall.modules.datasets.api.storage_location.schema import DatasetStorageLocation
+from dataall.modules.datasets.api.storage_location.types import DatasetStorageLocation
 
 createDatasetStorageLocation = gql.MutationField(
     name='createDatasetStorageLocation',
